improved module system and displaying denyprob plot in GUI
