+++ src/interface/mod.rs
@@ -1,0 +1,3 @@
+pub mod build_ui;
+pub mod detail;
+pub mod plotting;
